+++ ci-scripts/test/integration/pingcloud-services/prerequisites/00-test-deployment.sh
@@ -1,0 +1,56 @@
+#!/bin/bash
+
+. "${PROJECT_DIR}"/ci-scripts/common.sh "${1}"
+
+if skipTest "${0}"; then
+  log "Skipping test ${0}"
+  exit 0
+fi
+
+testP14COAuthDeploymentAvailability() {
+
+  status=$(kubectl get deployment p14c-oauth-service -n ${NAMESPACE} -o json | jq -r '.status.conditions[0].type')
+  assertEquals 0 $?
+  assertEquals "The status of the p14c-oauth-service deployment should be Available but was: ${status}" 'Available' ${status}
+}
+
+testP14CBootstrapDeploymentAvailability() {
+
+  status=$(kubectl get deployment p14c-bootstrap -n ${NAMESPACE} -o json | jq -r '.status.conditions[0].type')
+  assertEquals 0 $?
+  assertEquals "The status of the p14c-bootstrap deployment should be Available but was: ${status}" 'Available' ${status}
+}
+
+testP14CBOMDeploymentAvailability() {
+  status=$(kubectl get deployment p14c-bom-service -n ${NAMESPACE} -o json| jq -r '.status.conditions[0].type')
+  assertEquals 0 $?
+  assertEquals "The status of the p14c-bom-service deployment should be Available but was: ${status}" 'Available' ${status}
+}  
+
+testMetadataAvailability() {
+
+  exit_code=0
+  for i in {1..10}
+  do
+    testUrlsWithoutBasicAuthExpect2xx "${PINGCLOUD_METADATA}"
+    exit_code=$?
+
+    if [[ $exit_code -ne 0 ]]; then
+      log "The Pingcloud-metadata endpoint is inaccessible.  This is attempt ${i} of 10.  Wait 60 seconds and then try again..."
+      sleep 60
+    else
+      break
+    fi
+  done
+
+  assertEquals 0 $exit_code
+}
+
+# When arguments are passed to a script you must
+# consume all of them before shunit is invoked
+# or your script won't run.  For integration
+# tests, you need this line.
+shift $#
+
+# load shunit
+. ${SHUNIT_PATH}