+++ profiles/aws/pingdelegator/hooks/80-post-start.sh
@@ -1,4 +1,0 @@
-#!/usr/bin/env sh
-
-. "${HOOKS_DIR}/pingcommon.lib.sh"
-. "${HOOKS_DIR}/utils.lib.sh"